codes/spider/codes/requirements.txt
scrapy>=2.6.0
scrapy-redis>=0.7.0
pymongo>=4.0.0
redis>=4.0.0
pybloom-live>=3.0.0
requests>=2.27.0
kubernetes>=24.2.0
prometheus-client>=0.14.0
flask>=2.0.0
flask-cors>=3.0.10
pydantic>=1.10.0
aiohttp>=3.8.3
kafka-python>=2.0.2
elasticsearch>=8.6.0
PyJWT>=2.6.0
bcrypt>=4.0.1
python-dateutil>=2.8.2
gunicorn>=20.1.0
flask_limiter>=3.3.0
python-dotenv>=1.0.0 